Recover old nucliadb_utils nats
